loading="eager"
I experimented with transitioning between optimizeSpeed and then once
the page had loaded, transitioning to optimizeQuality.

However, this was littered with bugs like resizing the window and a lot
of assumptions like > 1920px should use optimizeQuality.

Out of this experimentation, the only thing that is being carried
through is loading="eager"